backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
google-generativeai>=0.3.0
python-multipart>=0.0.6
bcrypt>=4.0.1
